+++ apps/photos/src/components/Collections/CollectionShare/ShareControl/CollaboratorParticipants.tsx
@@ -5,7 +5,7 @@ import { t } from 'i18next';
 import { Collection } from 'types/collection';
 import { MenuItemGroup } from 'components/Menu/MenuItemGroup';
 import { EnteMenuItem } from 'components/Menu/EnteMenuItem';
-import AvatarCollectionShare from '../AvatarCollectionShare';
+import Avatar from 'components/pages/gallery/Avatar';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import MenuItemDivider from 'components/Menu/MenuItemDivider';
 import ManageAddCollab from './MangeAddCollab';
@@ -62,7 +62,7 @@ export function CollaboratorParticipants({
                             key={item}
                             onClick={() => openParticipantRoleView(item)}
                             label={item}
-                            startIcon={<AvatarCollectionShare email={item} />}
+                            startIcon={<Avatar email={item} />}
                             endIcon={<ChevronRightIcon />}
                         />
                         {index !== collaborators.length - 1 && (
