+++ apps/photos/src/components/pages/gallery/Avatar.tsx
@@ -6,7 +6,8 @@ import { useTheme } from '@mui/material/styles';
 import { logError } from 'utils/sentry';
 
 interface AvatarProps {
-    file: EnteFile;
+    file?: EnteFile;
+    email?: string;
 }
 
 const PUBLIC_COLLECTED_FILES_AVATAR_COLOR_CODE = '#000000';
@@ -24,7 +25,7 @@ const AvatarBase = styled('div')<{ colorCode: string; size: number }>`
     font-size: ${({ size }) => `${Math.floor(size / 2)}px`};
 `;
 
-const Avatar: React.FC<AvatarProps> = ({ file }) => {
+const Avatar: React.FC<AvatarProps> = ({ file, email }) => {
     const { userIDToEmailMap, user } = useContext(GalleryContext);
     const theme = useTheme();
 
@@ -58,9 +59,36 @@ const Avatar: React.FC<AvatarProps> = ({ file }) => {
                 setColorCode(PUBLIC_COLLECTED_FILES_AVATAR_COLOR_CODE);
             }
         } catch (err) {
-            logError(err, 'AvatarIcon.tsx - useLayoutEffect failed');
+            logError(err, 'AvatarIcon.tsx - useLayoutEffect file failed');
         }
-    }, []);
+    }, [file]);
+
+    useLayoutEffect(() => {
+        try {
+            if (!email) {
+                logError(Error(), 'email not found in userIDToEmailMap');
+                return;
+            }
+            const id = Array.from(userIDToEmailMap.keys()).find(
+                (key) => userIDToEmailMap.get(key) === email
+            );
+            if (!id && user.email !== email) {
+                logError(Error(), `ID not found for email: ${email}`);
+                return;
+            } else if (!id && user.email === email) {
+                setUserLetter(email[0].toUpperCase());
+                setColorCode(PUBLIC_COLLECTED_FILES_AVATAR_COLOR_CODE);
+                return;
+            }
+
+            const colorIndex = id % theme.colors.avatarColors.length;
+            const colorCode = theme.colors.avatarColors[colorIndex];
+            setUserLetter(email[0].toUpperCase());
+            setColorCode(colorCode);
+        } catch (err) {
+            logError(err, 'AvatarIcon.tsx - useLayoutEffect email failed');
+        }
+    }, [email]);
 
     if (!colorCode || !userLetter) {
         return <></>;
